refactor(services): implementar excepciones y NotificationService
Detalles de los cambios:
- Se implementaron query methods personalizados para NotificationRepository y el CRUD básico para NotificationService
- Para Tag/TrackingLis Service, se corrigieron los métodos para actualizar/eliminar

+++ src/main/java/com/example/studystreak/service/TrackingLinkService.java
@@ -88,8 +88,14 @@ public class TrackingLinkService {
         TrackingLink updatedtrackingLink = trackingLinkRepository.save(trackingLink);
         return modelMapper.map(updatedtrackingLink,TrackingLinkDTO.class);
     }
-    public void deleteTrackingLink(Long trackingId) {
+    public void deleteTrackingLink(Long trackingId, Long userId) {
         TrackingLink trackingLink = trackingLinkRepository.findById(trackingId).orElseThrow();
+
+        if (!trackingLink.getRequester().getId().equals(userId)
+                || !trackingLink.getReceiver().getId().equals(userId)) {
+            throw new RuntimeException();
+        }
+
         trackingLinkRepository.delete(trackingLink);
     }
 }
